Fix -Wsign-compare

+++ src/ray/common/placement_group.h
@@ -71,7 +71,7 @@ class PlacementGroupSpecBuilder {
     message_->set_placement_group_id(placement_group_id.Binary());
     message_->set_name(name);
     message_->set_strategy(strategy);
-    for (int i = 0; i < bundles.size(); i++) {
+    for (size_t i = 0; i < bundles.size(); i++) {
       auto resources = bundles[i];
       auto message_bundle = message_->add_bundles();
       auto mutable_bundle_id = message_bundle->mutable_bundle_id();
